:recycle: api: move CertificationAssessmentScore to src
