+++ api/src/certification/scoring/domain/models/CertificationAssessmentScore.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { status } from '../../../src/shared/domain/models/AssessmentResult.js';
+import { status } from '../../../../shared/domain/models/AssessmentResult.js';
 
 class CertificationAssessmentScore {
   constructor({
